feat(KONFLUX-9835): When repo got moved, that means it was deleted
Some info from Inderpreet:

In the past (I think before GitLab 16.0), deleting a project was a much
more immediate and final action. The problem with this approach was the
risk of accidental data loss. Restoring a deleted project often required
a complex and time-consuming process. In later versions Gitlab added a
kind of grace-period to keep the project with different name for some
days so that user can recover it in case it was an accidental deletion.
I think how it works now is as below:

First Deletion: Navigate to your project's Settings > General >
Advanced. Under the "Delete project" section, click the Delete project
button. This will change the project's status to "pending deletion" and
rename the repository to allow you to immediately create a new project
with the same name.

Second Deletion: After the first step, the "Delete project" button will
become available again. Clicking it a second time and confirming will
permanently and immediately delete the project.